requirements.txt
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
wheel
ipython
pyyaml
